api/requirements_qwen.txt
# Requirements for running Qwen domain suggestion model
torch>=2.0.0
transformers>=4.40.0
mlflow>=2.10.0
pandas>=1.5.0
requests>=2.31.0
python-dotenv>=1.0.0
fastapi>=0.68.0
uvicorn>=0.15.0
pydantic>=1.8.0